+++ .eslintrc.js
@@ -4,12 +4,13 @@ module.exports = {
     es2021: true,
     node: true,
   },
-  extends: ['standard'],
+  extends: ["standard"],
   parserOptions: {
     ecmaVersion: 12,
   },
   rules: {
-    'comma-dangle': 'off',
-    'space-before-function-paren': 'off',
+    "comma-dangle": "off",
+    "space-before-function-paren": "off",
+    "semi": "off",
   },
-}
+};
