+++ model/contacts.json
@@ -26,7 +26,7 @@
   {
     "id": 5,
     "name": "Cyrus Jackson",
-    "email": "nibh@semsempererat.com",
+    "email": "dui.hui.diam@Donec.com",
     "phone": "(501) 472-5218"
   },
   {
@@ -46,18 +46,5 @@
     "name": "Simon Morton",
     "email": "dui.Fusce.diam@Donec.com",
     "phone": "(233) 738-2360"
-  },
-  {
-    "id": 9,
-    "name": "Thomas Lucas",
-    "email": "nec@Nulla.com",
-    "phone": "(704) 398-7993"
-  },
-  {
-    "id": 10,
-    "name": "Alec Howard",
-    "email": "Donec.elementum@scelerisquescelerisquedui.net",
-    "phone": "(748) 206-2688"
   }
 ]
-
