+++ package.json
@@ -12,6 +12,7 @@
     "cors": "2.8.5",
     "cross-env": "7.0.3",
     "express": "4.17.1",
+    "joi": "^17.4.0",
     "morgan": "1.10.0",
     "uuid": "^8.3.2"
   },
